net: Move HttpSession to http
Relates to #1078
