+++ pyatv/support/net.py
@@ -7,19 +7,11 @@ import socket
 import struct
 from typing import List, Optional
 
-from aiohttp import ClientSession
 import netifaces
 
 from pyatv import exceptions
-from pyatv.support import log_binary
 
 _LOGGER = logging.getLogger(__name__)
-
-
-# This timeout is rather long and that is for a reason. If a device is sleeping, it
-# automatically wakes up when a service is requested from it. Up to 20 seconds or so
-# have been seen. So to deal with that, keep this high.
-DEFAULT_TIMEOUT = 25.0  # Seconds
 
 
 def unused_port() -> int:
@@ -121,75 +113,3 @@ def tcp_keepalive(sock) -> None:
             )
 
     _LOGGER.debug("Configured keep-alive on %s (%s)", sock, current_platform)
-
-
-class HttpSession:
-    """This class simplifies GET/POST requests."""
-
-    def __init__(self, client_session: ClientSession, base_url: str):
-        """Initialize a new HttpSession."""
-        self._session = client_session
-        self.base_url = base_url
-
-    async def get_data(self, path, headers=None, timeout=None):
-        """Perform a GET request."""
-        url = self.base_url + path
-        _LOGGER.debug("GET URL: %s", url)
-        resp = None
-        try:
-            resp = await self._session.get(
-                url,
-                headers=headers,
-                timeout=DEFAULT_TIMEOUT if timeout is None else timeout,
-            )
-            _LOGGER.debug(
-                "Response: status=%d, headers=[%s]",
-                resp.status,
-                ", ".join([f"{key}={value}" for key, value in resp.headers.items()]),
-            )
-            if resp.content_length is not None:
-                resp_data = await resp.read()
-                log_binary(_LOGGER, "<< GET", Data=resp_data)
-            else:
-                resp_data = None
-            return resp_data, resp.status
-        except Exception as ex:
-            if resp is not None:
-                resp.close()
-            raise ex
-        finally:
-            if resp is not None:
-                await resp.release()
-
-    async def post_data(self, path, data=None, headers=None, timeout=None):
-        """Perform a POST request."""
-        url = self.base_url + path
-        _LOGGER.debug("POST URL: %s", url)
-        log_binary(_LOGGER, ">> POST", Data=data)
-
-        resp = None
-        try:
-            resp = await self._session.post(
-                url,
-                headers=headers,
-                data=data,
-                timeout=DEFAULT_TIMEOUT if timeout is None else timeout,
-            )
-            _LOGGER.debug(
-                "Response: status=%d, headers=[%s]",
-                resp.status,
-                ", ".join([f"{key}={value}" for key, value in resp.headers.items()]),
-            )
-            if resp.content_length is not None:
-                resp_data = await resp.read()
-            else:
-                resp_data = None
-                log_binary(_LOGGER, "<< POST", Data=resp_data)
-            return resp_data, resp.status
-        except Exception as ex:
-            if resp is not None:
-                resp.close()
-            raise ex
-        finally:
-            if resp is not None:
-                await resp.release()
